New Quest Type: PickUp
implimented a new type of quest. A quest that can be fullfilled when gathering the needed Item! (Or the needed amount).
Will require no further changes than just adding 2 more coloms to the end of each row of RoomQuests - 1 for Amount (which we actually had forgotten had to be there) and 1 for a String type.
The Quest constructor will determine what that String is in terms of the Enum QuestType